pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrubfile"
version = "1.0.0"
description = "Scrub PII from PDFs, images, and DOCX files. Local-only. One command."
requires-python = ">=3.10"
dependencies = [
    "PyMuPDF>=1.24.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "easyocr>=1.7",
    "Pillow>=10.0",
    "python-docx>=1.0",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
]

[project.scripts]
scrubfile = "scrubfile.cli:app"

[project.optional-dependencies]
tesseract = ["pytesseract"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/scrubfile"]

[tool.pytest.ini_options]
testpaths = ["tests"]
